requirements.txt
asgiref==3.8.1
Django==5.1.4
python-barcode==0.15.1
sqlparse==0.5.2
